// src/app/enums/entity-type.enum.ts
// TODO EEIBKY - try to remove this enum if possible
export enum EntityType {
  FP = 'FeaturePacks',
  JBS = 'Jobs',
  SCHEDULES = "Schedules",
}

export enum EntityTranslateName {
  FeaturePack = 'featurePack',
  Job = 'job',
  Schedule = 'schedule'
}

export enum ListType {
  'MENU' = 'navigation.MENU',
}